style(utils): norm対応。error.c内の各種print関数をerror_print.cに移動
print_errorは幅広く使うのでerror.cに残しています
