+++ Makefile
@@ -6,7 +6,7 @@
 #    By: nfukada <nfukada@student.42tokyo.jp>       +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2021/02/10 18:36:52 by nfukada           #+#    #+#              #
-#    Updated: 2021/03/03 21:56:21 by nfukada          ###   ########.fr        #
+#    Updated: 2021/03/05 18:03:45 by nfukada          ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -23,14 +23,14 @@ LIBFT_DIR	:= libft
 LIBFT		:= $(LIBFT_DIR)/libft.a
 INC			:= -I$(INC_DIR) -I$(LIBFT_DIR)
 
-SRCS		:= srcs/lexer/lexer_general_state.c srcs/lexer/lexer.c srcs/lexer/lexer_quote_state.c srcs/utils/shell_init.c srcs/utils/shlvl_init.c srcs/utils/token_utils.c srcs/utils/token.c srcs/utils/env_utils_get.c srcs/utils/env.c srcs/utils/error.c srcs/utils/signal.c srcs/utils/env_utils.c srcs/parser/parser_utils.c srcs/parser/redirect.c srcs/parser/print_node_label.c srcs/parser/parser.c srcs/parser/node.c srcs/parser/print_node.c srcs/expander/expander.c srcs/expander/expand_tokens.c srcs/expander/expand_var.c srcs/main.c srcs/exec/exec_command.c srcs/exec/exec.c srcs/exec/build_path_stat.c srcs/exec/command.c srcs/exec/build_path.c srcs/exec/exec_pipe.c srcs/exec/exec_redirect.c srcs/exec/build_path_utils.c srcs/builtin/mini_echo.c srcs/builtin/builtin.c srcs/builtin/mini_exit.c 
+SRCS		:= srcs/lexer/lexer_general_state.c srcs/lexer/lexer.c srcs/lexer/lexer_quote_state.c srcs/utils/shell_init.c srcs/utils/shlvl_init.c srcs/utils/token_utils.c srcs/utils/token.c srcs/utils/env_utils_get.c srcs/utils/env.c srcs/utils/error.c srcs/utils/error_print.c srcs/utils/signal.c srcs/utils/env_utils.c srcs/parser/parser_utils.c srcs/parser/redirect.c srcs/parser/print_node_label.c srcs/parser/parser.c srcs/parser/node.c srcs/parser/print_node.c srcs/expander/expander.c srcs/expander/expand_tokens.c srcs/expander/expand_var.c srcs/main.c srcs/exec/exec_command.c srcs/exec/exec.c srcs/exec/build_path_stat.c srcs/exec/command.c srcs/exec/build_path.c srcs/exec/exec_pipe.c srcs/exec/exec_redirect.c srcs/exec/build_path_utils.c srcs/builtin/mini_echo.c srcs/builtin/builtin.c srcs/builtin/mini_exit.c 
 OBJS		:= $(SRCS:%.c=%.o)
 LIBS		:= -lft -L$(LIBFT_DIR)
 
 DEBUG		:= FALSE
 
 CC			:= gcc
-CFLAGS		:= -Wall -Werror -Wextra $(INC) -g -fsanitize=address -D DEBUG=$(DEBUG)
+CFLAGS		:= -Wall -Werror -Wextra $(INC) -D DEBUG=$(DEBUG)
 
 .PHONY: all clean fclean re bonus debug norm srcs
